+++ src/components/NavBar/Logo.jsx
@@ -1,0 +1,12 @@
+function Logo({ onClick }) {
+  return (
+    <div
+      className="text-4xl sm:text-5xl font-bold cursor-pointer"
+      onClick={onClick}
+    >
+      🎬 Pickflix
+    </div>
+  );
+}
+
+export default Logo;
